TaskSGP4 receives TLEs and sends locations and velocities! Needs testing!

+++ Src/sgp4_tle.cpp
@@ -40,16 +40,16 @@ namespace sgp4_utils
         // Line 1 Parsing
         tle.satelliteNumber = std::stoi(l1.substr(2, 5));               // Columns 3-7
         tle.epochYear = std::stoi(l1.substr(18, 2));                    // Columns 19-20
-        tle.epochDay = std::stod(l1.substr(20, 12));                    // Columns 21-32
-        tle.meanMotionDerivative1 = std::stod(l1.substr(33, 10));       // Columns 34-43
-        tle.meanMotionDerivative2 = std::stod("0." + l1.substr(44, 8)); // Columns 45-52 (decimal point assumed)
+        tle.epochDay = std::stof(l1.substr(20, 12));                    // Columns 21-32
+        tle.meanMotionDerivative1 = std::stof(l1.substr(33, 10));       // Columns 34-43
+        tle.meanMotionDerivative2 = std::stof("0." + l1.substr(44, 8)); // Columns 45-52 (decimal point assumed)
 
         // Handle the exponent form of bstar
         std::string bStarStr = l1.substr(53, 8);
         std::string bStarSign = bStarStr.substr(0, 1);
         std::string bStarInteger = bStarStr.substr(1, 5);
         std::string bStarExponent = bStarStr.substr(6, 2);
-        tle.bStarDrag = std::stod(bStarSign + "0." + bStarInteger + "e" + bStarExponent);
+        tle.bStarDrag = std::stof(bStarSign + "0." + bStarInteger + "e" + bStarExponent);
 
         tle.ephemerisType = std::stoi(l1.substr(62, 1));                // Column 63
         tle.elementNumber = std::stoi(l1.substr(64, 4));                // Columns 65-68
@@ -57,12 +57,12 @@ namespace sgp4_utils
             return std::nullopt;
 
         // Line 2 Parsing
-        tle.inclination = std::stod(l2.substr(8, 8));                   // Columns 9-16
-        tle.rightAscensionAscendingNode = std::stod(l2.substr(17, 8));  // Columns 18-25
-        tle.eccentricity = std::stod("0." + l2.substr(26, 7));          // Columns 27-33 (decimal point assumed)
-        tle.argumentOfPerigee = std::stod(l2.substr(34, 8));            // Columns 35-42
-        tle.meanAnomaly = std::stod(l2.substr(43, 8));                  // Columns 44-51
-        tle.meanMotion = std::stod(l2.substr(52, 11));                  // Columns 53-63
+        tle.inclination = std::stof(l2.substr(8, 8));                   // Columns 9-16
+        tle.rightAscensionAscendingNode = std::stof(l2.substr(17, 8));  // Columns 18-25
+        tle.eccentricity = std::stof("0." + l2.substr(26, 7));          // Columns 27-33 (decimal point assumed)
+        tle.argumentOfPerigee = std::stof(l2.substr(34, 8));            // Columns 35-42
+        tle.meanAnomaly = std::stof(l2.substr(43, 8));                  // Columns 44-51
+        tle.meanMotion = std::stof(l2.substr(52, 11));                  // Columns 53-63
         tle.revolutionNumberAtEpoch = std::stoi(l2.substr(63, 5));      // Columns 64-68
         if (l2[68] != checksum(l2))
             return std::nullopt;
